requirements.txt
sqlalchemy
RPi.GPIO
flask
Flask-Bootstrap
Flask-WTF
